+++ src/pages/_app.tsx
@@ -3,6 +3,7 @@ import "../components/layout.css";
 import { ChakraProvider, CSSReset, Skeleton } from "@chakra-ui/core";
 import customTheme from "../theme";
 import Navigation from "../components/organisms/navigation";
+import Head from "next/head";
 import Layout from "../components/layout";
 import { useFetchSession } from "../utils/user";
 import { isLeft, isRight } from "fp-ts/lib/Either";
@@ -19,6 +20,16 @@ function MyApp({ Component, pageProps }: AppProps) {
 
   return (
     <>
+      <Head>
+        <script
+          dangerouslySetInnerHTML={{
+            __html:
+              isRight(sessionAndThunk[0]) && isRight(sessionAndThunk[0].right)
+                ? `window.intercomSettings = { app_id: "${process.env.INTERCOM_ID}", email: "${sessionAndThunk[0].right.right.user.email}"};`
+                : `window.intercomSettings = { app_id: "${process.env.INTERCOM_ID}"};`,
+          }}
+        ></script>
+      </Head>
       <ChakraProvider theme={customTheme}>
         <CSSReset />
         <Layout>
